refactor: extract JsonRpcService into dedicated module
- Move JsonRpcService from router.rs to new src/jsonrpc.rs
- Separates JSON-RPC 2.0 framing concerns from MCP routing
- Reduces router.rs from 2368 to 2120 lines
- JsonRpcService can now wrap any MCP service, not just McpRouter
- Update imports in lib.rs, stdio.rs, and http.rs
