+++ src/router.rs
@@ -863,266 +863,10 @@ impl Service<RouterRequest> for McpRouter {
     }
 }
 
-// =============================================================================
-// JSON-RPC layer
-// =============================================================================
-
-/// Service that handles JSON-RPC framing
-///
-/// Wraps an McpRouter and handles JSON-RPC request/response conversion.
-pub struct JsonRpcService<S> {
-    inner: S,
-}
-
-impl<S> JsonRpcService<S> {
-    pub fn new(inner: S) -> Self {
-        Self { inner }
-    }
-
-    /// Process a single JSON-RPC request
-    pub async fn call_single(&mut self, req: JsonRpcRequest) -> Result<JsonRpcResponse>
-    where
-        S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-            + Clone
-            + Send
-            + 'static,
-        S::Future: Send,
-    {
-        process_single_request(self.inner.clone(), req).await
-    }
-
-    /// Process a batch of JSON-RPC requests concurrently
-    pub async fn call_batch(
-        &mut self,
-        requests: Vec<JsonRpcRequest>,
-    ) -> Result<Vec<JsonRpcResponse>>
-    where
-        S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-            + Clone
-            + Send
-            + 'static,
-        S::Future: Send,
-    {
-        if requests.is_empty() {
-            return Err(Error::JsonRpc(JsonRpcError::invalid_request(
-                "Empty batch request",
-            )));
-        }
-
-        // Process all requests concurrently
-        let futures: Vec<_> = requests
-            .into_iter()
-            .map(|req| {
-                let inner = self.inner.clone();
-                let req_id = req.id.clone();
-                async move {
-                    match process_single_request(inner, req).await {
-                        Ok(resp) => resp,
-                        Err(e) => {
-                            // Convert errors to error responses instead of dropping
-                            JsonRpcResponse::error(
-                                Some(req_id),
-                                JsonRpcError::internal_error(e.to_string()),
-                            )
-                        }
-                    }
-                }
-            })
-            .collect();
-
-        let results: Vec<JsonRpcResponse> = futures::future::join_all(futures).await;
-
-        // Results will never be empty since we converted all errors to responses
-        Ok(results)
-    }
-
-    /// Process a JSON-RPC message (single or batch)
-    pub async fn call_message(&mut self, msg: JsonRpcMessage) -> Result<JsonRpcResponseMessage>
-    where
-        S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-            + Clone
-            + Send
-            + 'static,
-        S::Future: Send,
-    {
-        match msg {
-            JsonRpcMessage::Single(req) => {
-                let response = self.call_single(req).await?;
-                Ok(JsonRpcResponseMessage::Single(response))
-            }
-            JsonRpcMessage::Batch(requests) => {
-                let responses = self.call_batch(requests).await?;
-                Ok(JsonRpcResponseMessage::Batch(responses))
-            }
-        }
-    }
-}
-
-impl<S> Clone for JsonRpcService<S>
-where
-    S: Clone,
-{
-    fn clone(&self) -> Self {
-        Self {
-            inner: self.inner.clone(),
-        }
-    }
-}
-
-impl<S> Service<JsonRpcRequest> for JsonRpcService<S>
-where
-    S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-        + Clone
-        + Send
-        + 'static,
-    S::Future: Send,
-{
-    type Response = JsonRpcResponse;
-    type Error = Error;
-    type Future =
-        Pin<Box<dyn Future<Output = std::result::Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<std::result::Result<(), Self::Error>> {
-        self.inner.poll_ready(cx).map_err(|_| unreachable!())
-    }
-
-    fn call(&mut self, req: JsonRpcRequest) -> Self::Future {
-        let mut inner = self.inner.clone();
-        Box::pin(async move {
-            // Parse the MCP request from JSON-RPC
-            let mcp_request = McpRequest::from_jsonrpc(&req)?;
-
-            // Create router request
-            let router_req = RouterRequest {
-                id: req.id,
-                inner: mcp_request,
-            };
-
-            // Call the inner service
-            let response = inner.call(router_req).await.unwrap(); // Infallible
-
-            // Convert to JSON-RPC response
-            Ok(response.into_jsonrpc())
-        })
-    }
-}
-
-/// Service implementation for JSON-RPC batch requests
-impl<S> Service<JsonRpcMessage> for JsonRpcService<S>
-where
-    S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-        + Clone
-        + Send
-        + 'static,
-    S::Future: Send,
-{
-    type Response = JsonRpcResponseMessage;
-    type Error = Error;
-    type Future =
-        Pin<Box<dyn Future<Output = std::result::Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<std::result::Result<(), Self::Error>> {
-        self.inner.poll_ready(cx).map_err(|_| unreachable!())
-    }
-
-    fn call(&mut self, msg: JsonRpcMessage) -> Self::Future {
-        let inner = self.inner.clone();
-        Box::pin(async move {
-            match msg {
-                JsonRpcMessage::Single(req) => {
-                    let response = process_single_request(inner, req).await?;
-                    Ok(JsonRpcResponseMessage::Single(response))
-                }
-                JsonRpcMessage::Batch(requests) => {
-                    if requests.is_empty() {
-                        // Empty batch is an invalid request per JSON-RPC spec
-                        return Ok(JsonRpcResponseMessage::Single(JsonRpcResponse::error(
-                            None,
-                            JsonRpcError::invalid_request("Empty batch request"),
-                        )));
-                    }
-
-                    // Process all requests concurrently
-                    let futures: Vec<_> = requests
-                        .into_iter()
-                        .map(|req| {
-                            let inner = inner.clone();
-                            let req_id = req.id.clone();
-                            async move {
-                                match process_single_request(inner, req).await {
-                                    Ok(resp) => resp,
-                                    Err(e) => {
-                                        // Convert errors to error responses instead of dropping
-                                        JsonRpcResponse::error(
-                                            Some(req_id),
-                                            JsonRpcError::internal_error(e.to_string()),
-                                        )
-                                    }
-                                }
-                            }
-                        })
-                        .collect();
-
-                    let results: Vec<JsonRpcResponse> = futures::future::join_all(futures).await;
-
-                    // Empty results only possible if input was empty (already handled above)
-                    if results.is_empty() {
-                        return Ok(JsonRpcResponseMessage::Single(JsonRpcResponse::error(
-                            None,
-                            JsonRpcError::internal_error("All batch requests failed"),
-                        )));
-                    }
-
-                    Ok(JsonRpcResponseMessage::Batch(results))
-                }
-            }
-        })
-    }
-}
-
-/// Helper function to process a single JSON-RPC request
-async fn process_single_request<S>(
-    mut inner: S,
-    req: JsonRpcRequest,
-) -> std::result::Result<JsonRpcResponse, Error>
-where
-    S: Service<RouterRequest, Response = RouterResponse, Error = std::convert::Infallible>
-        + Send
-        + 'static,
-    S::Future: Send,
-{
-    // Validate JSON-RPC version
-    if let Err(e) = req.validate() {
-        return Ok(JsonRpcResponse::error(Some(req.id), e));
-    }
-
-    // Parse the MCP request from JSON-RPC
-    let mcp_request = match McpRequest::from_jsonrpc(&req) {
-        Ok(r) => r,
-        Err(e) => {
-            return Ok(JsonRpcResponse::error(
-                Some(req.id),
-                JsonRpcError::invalid_params(e.to_string()),
-            ));
-        }
-    };
-
-    // Create router request
-    let router_req = RouterRequest {
-        id: req.id,
-        inner: mcp_request,
-    };
-
-    // Call the inner service
-    let response = inner.call(router_req).await.unwrap(); // Infallible
-
-    // Convert to JSON-RPC response
-    Ok(response.into_jsonrpc())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::jsonrpc::JsonRpcService;
     use crate::tool::ToolBuilder;
     use schemars::JsonSchema;
     use serde::Deserialize;
